Apply similar changes to other rulesets
