+++ osu.Game.Rulesets.Taiko/Difficulty/TaikoDifficultyCalculator.cs
@@ -1,22 +1,20 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Game.Beatmaps;
-using osu.Game.Rulesets.Mania.Beatmaps;
-using osu.Game.Rulesets.Mania.Objects;
-using osu.Game.Rulesets.Mods;
 using System;
 using System.Collections.Generic;
+using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Difficulty;
+using osu.Game.Rulesets.Taiko.Objects;
 
-namespace osu.Game.Rulesets.Mania
+namespace osu.Game.Rulesets.Taiko.Difficulty
 {
-    internal class ManiaDifficultyCalculator : DifficultyCalculator
+    internal class TaikoDifficultyCalculator : DifficultyCalculator
     {
-        private const double star_scaling_factor = 0.018;
+        private const double star_scaling_factor = 0.04125;
 
         /// <summary>
-        /// In milliseconds. For difficulty calculation we will only look at the highest strain value in each time interval of size strain_step.
+        /// In milliseconds. For difficulty calculation we will only look at the highest strain value in each time interval of size STRAIN_STEP.
         /// This is to eliminate higher influence of stream over aim by simply having more HitObjects with high strain.
         /// The higher this value, the less strains there will be, indirectly giving long beatmaps an advantage.
         /// </summary>
@@ -30,15 +28,10 @@ namespace osu.Game.Rulesets.Mania
         /// <summary>
         /// HitObjects are stored as a member variable.
         /// </summary>
-        private readonly List<ManiaHitObjectDifficulty> difficultyHitObjects = new List<ManiaHitObjectDifficulty>();
+        private readonly List<TaikoHitObjectDifficulty> difficultyHitObjects = new List<TaikoHitObjectDifficulty>();
 
-        public ManiaDifficultyCalculator(IBeatmap beatmap)
+        public TaikoDifficultyCalculator(IBeatmap beatmap)
             : base(beatmap)
-        {
-        }
-
-        public ManiaDifficultyCalculator(IBeatmap beatmap, Mod[] mods)
-            : base(beatmap, mods)
         {
         }
 
@@ -47,20 +40,21 @@ namespace osu.Game.Rulesets.Mania
             // Fill our custom DifficultyHitObject class, that carries additional information
             difficultyHitObjects.Clear();
 
-            int columnCount = (Beatmap as ManiaBeatmap)?.TotalColumns ?? 7;
-
             foreach (var hitObject in Beatmap.HitObjects)
-                difficultyHitObjects.Add(new ManiaHitObjectDifficulty((ManiaHitObject)hitObject, columnCount));
+                difficultyHitObjects.Add(new TaikoHitObjectDifficulty((TaikoHitObject)hitObject));
 
             // Sort DifficultyHitObjects by StartTime of the HitObjects - just to make sure.
             difficultyHitObjects.Sort((a, b) => a.BaseHitObject.StartTime.CompareTo(b.BaseHitObject.StartTime));
 
-            if (!calculateStrainValues())
-                return 0;
+            if (!calculateStrainValues()) return 0;
 
             double starRating = calculateDifficulty() * star_scaling_factor;
 
-            categoryDifficulty?.Add("Strain", starRating);
+            if (categoryDifficulty != null)
+            {
+                categoryDifficulty.Add("Strain", starRating);
+                categoryDifficulty.Add("Hit window 300", 35 /*HitObjectManager.HitWindow300*/ / TimeRate);
+            }
 
             return starRating;
         }
@@ -68,12 +62,11 @@ namespace osu.Game.Rulesets.Mania
         private bool calculateStrainValues()
         {
             // Traverse hitObjects in pairs to calculate the strain value of NextHitObject from the strain value of CurrentHitObject and environment.
-            using (List<ManiaHitObjectDifficulty>.Enumerator hitObjectsEnumerator = difficultyHitObjects.GetEnumerator())
+            using (List<TaikoHitObjectDifficulty>.Enumerator hitObjectsEnumerator = difficultyHitObjects.GetEnumerator())
             {
-                if (!hitObjectsEnumerator.MoveNext())
-                    return false;
+                if (!hitObjectsEnumerator.MoveNext()) return false;
 
-                ManiaHitObjectDifficulty current = hitObjectsEnumerator.Current;
+                TaikoHitObjectDifficulty current = hitObjectsEnumerator.Current;
 
                 // First hitObject starts at strain 1. 1 is the default for strain values, so we don't need to set it here. See DifficultyHitObject.
                 while (hitObjectsEnumerator.MoveNext())
@@ -96,7 +89,7 @@ namespace osu.Game.Rulesets.Mania
             double intervalEndTime = actualStrainStep;
             double maximumStrain = 0; // We need to keep track of the maximum strain in the current interval
 
-            ManiaHitObjectDifficulty previousHitObject = null;
+            TaikoHitObjectDifficulty previousHitObject = null;
             foreach (var hitObject in difficultyHitObjects)
             {
                 // While we are beyond the current interval push the currently available maximum to our strain list
@@ -112,9 +105,8 @@ namespace osu.Game.Rulesets.Mania
                     }
                     else
                     {
-                        double individualDecay = Math.Pow(ManiaHitObjectDifficulty.INDIVIDUAL_DECAY_BASE, (intervalEndTime - previousHitObject.BaseHitObject.StartTime) / 1000);
-                        double overallDecay = Math.Pow(ManiaHitObjectDifficulty.OVERALL_DECAY_BASE, (intervalEndTime - previousHitObject.BaseHitObject.StartTime) / 1000);
-                        maximumStrain = previousHitObject.IndividualStrain * individualDecay + previousHitObject.OverallStrain * overallDecay;
+                        double decay = Math.Pow(TaikoHitObjectDifficulty.DECAY_BASE, (intervalEndTime - previousHitObject.BaseHitObject.StartTime) / 1000);
+                        maximumStrain = previousHitObject.Strain * decay;
                     }
 
                     // Go to the next time interval
@@ -122,8 +114,7 @@ namespace osu.Game.Rulesets.Mania
                 }
 
                 // Obtain maximum strain
-                double strain = hitObject.IndividualStrain + hitObject.OverallStrain;
-                maximumStrain = Math.Max(strain, maximumStrain);
+                maximumStrain = Math.Max(hitObject.Strain, maximumStrain);
 
                 previousHitObject = hitObject;
             }
